Change restaurant id to Long

+++ api_server/src/main/java/io/zetch/app/repo/RestaurantRepository.java
@@ -3,5 +3,5 @@ package io.zetch.app.repo;
 import io.zetch.app.domain.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantRepository extends JpaRepository<Restaurant, Integer> {
+public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
 }
